Moved handling connector events to a hook
